add deleteAllWeeks function as part of redux and rename course_actions_redux as course_actions
