+++ app/assets/javascripts/constants/timeline.js
@@ -9,3 +9,4 @@ export const ADD_BLOCK = 'ADD_BLOCK';
 export const DELETE_BLOCK = 'DELETE_BLOCK';
 export const INSERT_BLOCK = 'INSERT_BLOCK';
 export const RESTORE_TIMELINE = 'RESTORE_TIMELINE';
+export const DELETE_ALL_WEEKS = 'DELETE_ALL_WEEKS';
